LayoutTests/fast/text/line-break-after-question-mark-expected.txt

Tests when line breaking is allowed after a question mark character.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS allowsBreakBefore(33) is false
PASS allowsBreakBefore(34) is false
PASS allowsBreakBefore(35) is true
PASS allowsBreakBefore(36) is true
PASS allowsBreakBefore(37) is true
PASS allowsBreakBefore(38) is true
PASS allowsBreakBefore(40) is true
PASS allowsBreakBefore(41) is false
PASS allowsBreakBefore(42) is true
PASS allowsBreakBefore(43) is true
PASS allowsBreakBefore(44) is false
PASS allowsBreakBefore(45) is true
PASS allowsBreakBefore(46) is false
PASS allowsBreakBefore(47) is false
PASS allowsBreakBefore(48) is true
PASS allowsBreakBefore(49) is true
PASS allowsBreakBefore(50) is true
PASS allowsBreakBefore(51) is true
PASS allowsBreakBefore(52) is true
PASS allowsBreakBefore(53) is true
PASS allowsBreakBefore(54) is true
PASS allowsBreakBefore(55) is true
PASS allowsBreakBefore(56) is true
PASS allowsBreakBefore(57) is true
PASS allowsBreakBefore(58) is false
PASS allowsBreakBefore(59) is false
PASS allowsBreakBefore(60) is true
PASS allowsBreakBefore(61) is true
PASS allowsBreakBefore(62) is true
PASS allowsBreakBefore(63) is false
PASS allowsBreakBefore(64) is true
PASS allowsBreakBefore(65) is true
PASS allowsBreakBefore(66) is true
PASS allowsBreakBefore(67) is true
PASS allowsBreakBefore(68) is true
PASS allowsBreakBefore(69) is true
PASS allowsBreakBefore(70) is true
PASS allowsBreakBefore(71) is true
PASS allowsBreakBefore(72) is true
PASS allowsBreakBefore(73) is true
PASS allowsBreakBefore(74) is true
PASS allowsBreakBefore(75) is true
PASS allowsBreakBefore(76) is true
PASS allowsBreakBefore(77) is true
PASS allowsBreakBefore(78) is true
PASS allowsBreakBefore(79) is true
PASS allowsBreakBefore(80) is true
PASS allowsBreakBefore(81) is true
PASS allowsBreakBefore(82) is true
PASS allowsBreakBefore(83) is true
PASS allowsBreakBefore(84) is true
PASS allowsBreakBefore(85) is true
PASS allowsBreakBefore(86) is true
PASS allowsBreakBefore(87) is true
PASS allowsBreakBefore(88) is true
PASS allowsBreakBefore(89) is true
PASS allowsBreakBefore(90) is true
PASS allowsBreakBefore(91) is true
PASS allowsBreakBefore(92) is true
PASS allowsBreakBefore(93) is false
PASS allowsBreakBefore(94) is true
PASS allowsBreakBefore(95) is true
PASS allowsBreakBefore(96) is true
PASS allowsBreakBefore(97) is true
PASS allowsBreakBefore(98) is true
PASS allowsBreakBefore(99) is true
PASS allowsBreakBefore(100) is true
PASS allowsBreakBefore(101) is true
PASS allowsBreakBefore(102) is true
PASS allowsBreakBefore(103) is true
PASS allowsBreakBefore(104) is true
PASS allowsBreakBefore(105) is true
PASS allowsBreakBefore(106) is true
PASS allowsBreakBefore(107) is true
PASS allowsBreakBefore(108) is true
PASS allowsBreakBefore(109) is true
PASS allowsBreakBefore(110) is true
PASS allowsBreakBefore(111) is true
PASS allowsBreakBefore(112) is true
PASS allowsBreakBefore(113) is true
PASS allowsBreakBefore(114) is true
PASS allowsBreakBefore(115) is true
PASS allowsBreakBefore(116) is true
PASS allowsBreakBefore(117) is true
PASS allowsBreakBefore(118) is true
PASS allowsBreakBefore(119) is true
PASS allowsBreakBefore(120) is true
PASS allowsBreakBefore(121) is true
PASS allowsBreakBefore(122) is true
PASS allowsBreakBefore(123) is true
PASS allowsBreakBefore(124) is true
PASS allowsBreakBefore(125) is false
PASS allowsBreakBefore(126) is true
PASS successfullyParsed is true

TEST COMPLETE
